refactor(Amount,BonusNumber): wrapper 패키지로 이동
