+++ src/main/java/lotto/controller/LottoController.java
@@ -1,7 +1,7 @@
 package lotto.controller;
 
-import lotto.domain.Amount;
-import lotto.domain.BonusNumber;
+import lotto.wrapper.Amount;
+import lotto.wrapper.BonusNumber;
 import lotto.domain.Lotto;
 import lotto.domain.Lottos;
 import lotto.service.LottoService;
