+++ src/main/java/lotto/service/LottoService.java
@@ -4,8 +4,8 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import lotto.domain.Amount;
-import lotto.domain.BonusNumber;
+import lotto.wrapper.Amount;
+import lotto.wrapper.BonusNumber;
 import lotto.domain.Lotto;
 import lotto.domain.Lottos;
 import lotto.util.LottoNumberGenerator;
