+++ src/main/java/lotto/wrapper/Amount.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.wrapper;
 
 public class Amount {
