+++ src/test/java/lotto/wrapper/BonusNumberTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.wrapper;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
